requirements.txt
PyYAML==6.0.1
jsonschema>=4.18,<5

# Optional: watchdog for efficient watch mode (fallback to polling if not installed)
# watchdog>=4.0.1
